move address fetcher to hyperdrive/addresses.py
