+++ src/bots/checkpoint_bot.py
@@ -11,6 +11,7 @@ from elfpy.utils import logs
 from eth_account.account import Account
 from web3.contract.contract import Contract
 
+import src.hyperdrive.addresses
 from src import eth, hyperdrive
 from src.eth.accounts.eth_account import EthAgent
 from src.eth.rpc_interface import set_anvil_account_balance
@@ -78,7 +79,7 @@ def main() -> None:
 
     # Get the Hyperdrive contract.
     hyperdrive_abis = eth.abi.load_all_abis(config.abi_folder)
-    addresses = hyperdrive.contract_interface.fetch_hyperdrive_address_from_url(
+    addresses = src.hyperdrive.addresses.fetch_hyperdrive_address_from_url(
         os.path.join(config.artifacts_url, "addresses.json")
     )
     hyperdrive_contract: Contract = web3.eth.contract(
