+++ src/data/acquire_data.py
@@ -13,6 +13,7 @@ from eth_utils import address
 from web3 import Web3
 from web3.contract.contract import Contract
 
+import src.hyperdrive.addresses
 from src import eth, hyperdrive
 from src.data import convert_data, postgres
 
@@ -60,7 +61,7 @@ def main(
 
     # send a request to the local server to fetch the deployed contract addresses and
     # all Hyperdrive contract addresses from the server response
-    addresses = hyperdrive.contract_interface.fetch_hyperdrive_address_from_url(contracts_url)
+    addresses = src.hyperdrive.addresses.fetch_hyperdrive_address_from_url(contracts_url)
     abis = eth.abi.load_all_abis(abi_dir)
 
     hyperdrive_contract = hyperdrive.contract_interface.get_hyperdrive_contract(web3, abis, addresses)
