+++ src/eth_bots/fund_bots.py
@@ -1,7 +1,5 @@
 """Fund agent private keys from a user key."""
-from __future__ import annotations
-
-import json
+from __future__ i..hyperdriveport annaddressesson
 import os
 
 from dotenv import load_dotenv
@@ -53,7 +51,7 @@ if __name__ == "__main__":
         environment_config.base_abi + ".json",
     )
     base_contract_abi = eth.abi.load_abi_from_file(abi_file_loc)
-    addresses = hyperdrive.contract_interface.fetch_hyperdrive_address_from_url(
+    addresses = addresses.fetch_hyperdrive_address_from_url(
         os.path.join(environment_config.artifacts_url, "addresses.json")
     )
     base_token_contract = web3.eth.contract(
